% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_unique}
\alias{db_unique}
\title{Handling of duplicates}
\usage{
db_unique(db, uuid = "uuid", MTIME = "MTIME")
}
\arguments{
\item{db}{data frame}

\item{uuid}{string denoting the attribute used as universal-unique-identifier. Defaults to 'uuid'}

\item{MTIME}{string denoting the attribute used as universal-unique-identifier. Defaults to 'MTIME'}
}
\description{
Identifies duplicated entries based on the uuid attribute and only returns the most recent entry, selected by the MTIME attribute
}
\examples{
## code example dataset
(db <- data.frame(uuid = LETTERS[1:10], MTIME = Sys.time()))
## duplicate last entry db <- rbind(db,db[10,])
## modify the time stamp by adding 1hour
db[11, "MTIME"] <- db[11, "MTIME"] + 60*60
db
## Handle duplicate using the db_unique function
(db <- db_unique(db))

}
